minimal-requirements.txt
# minimal-requirements.txt
fastapi==0.68.1
selenium==4.9.1
beautifulsoup4==4.9.3
pydantic==1.9.0
uvicorn==0.15.0
python-dotenv==0.19.0
aiohttp==3.8.1
PyYAML==6.0
undetected-chromedriver==3.4.6
requests==2.31.0
email-validator